requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.12.1
asgiref==3.11.0
attrs==25.4.0
autobahn==25.12.2
Automat==25.4.16
cbor2==5.8.0
certifi==2026.1.4
cffi==2.0.0
channels==4.3.2
channels_redis==4.3.0
charset-normalizer==3.4.4
colorama==0.4.6
constantly==23.10.4
cryptography==46.0.3
daphne==4.2.1
deepgram-sdk==5.3.1
Django==6.0.1
frozenlist==1.8.0
google-ai-generativelanguage==0.6.15
google-api-core==2.25.2
google-api-python-client==2.188.0
google-auth==2.48.0rc0
google-auth-httplib2==0.3.0
google-generativeai==0.8.6
googleapis-common-protos==1.72.0
grpcio==1.76.0
grpcio-status==1.71.2
h11==0.16.0
httpcore==1.0.9
httplib2==0.31.1
httpx==0.28.1
hyperlink==21.0.0
idna==3.11
Incremental==24.11.0
msgpack==1.1.2
multidict==6.7.0
packaging==26.0
propcache==0.4.1
proto-plus==1.27.0
protobuf==5.29.5
py-ubjson==0.16.1
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
pyOpenSSL==25.3.0
pyparsing==3.3.2
python-dotenv==1.2.1
redis==7.1.0
requests==2.32.5
rsa==4.9.1
service-identity==24.2.0
sqlparse==0.5.5
tqdm==4.67.1
Twisted==25.5.0
txaio==25.12.2
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
ujson==5.11.0
uritemplate==4.2.0
urllib3==2.6.3
websockets==16.0
yarl==1.22.0
yt-dlp==2025.12.8
zope.interface==8.2
